requirements.txt
webargs>=5.5.2
deepaas>=1.3.0
importlib-metadata>=1.7.0
aiohttp
pyyaml
